refactor: move scantree logic into helpers.js
